Fix UI issues: Tailwind class, dedupe loading, remove dead code
